Split transport session tests into dedicated module
